+++ data/sponsor.md
@@ -1,0 +1,73 @@
+# Sponsor
+
+Sponsoring EuroPython guarantees you highly targeted visibility and the
+opportunity to present yourself and your company to one of the largest and most
+diverse Python communities in Europe and beyond.
+
+As a sponsor of EuroPython 2022, you will directly help promote the work of the
+great open-source community and help further its development.
+
+# Sponsor Team
+
+If you are interested in sponsoring EuroPython, please send an email to our
+sponsor team at:
+
+- sponsoring@europython.eu
+
+# Sponsor Packages & Brochure
+
+We are still preparing the EuroPython 2022 packages. In the meantime, please
+have a look at our EuroPython 2019 Sponsor Brochure and the packages.
+
+- [EuroPython 2019 Sponsor Brochure](https://ep2019.europython.eu/sponsor/brochure/)
+  (in-person)
+- [EuroPython 2019 Sponsor Packages](https://ep2019.europython.eu/sponsor/packages/)
+  (in-person)
+
+The prices for 2022 haven’t been finalized yet, but will likely see an increase.
+As usual, we will also offer a 10% discount for early bird sponsors and free
+tables for Open Source Software.
+
+# Appendix
+
+## **Old 2020 Precovid Pre-launch Site: Sponsoring EuroPython 2020**
+
+Sponsoring EuroPython guarantees you highly targeted visibility and the
+opportunity to present yourself and your company in a professional and
+innovative environment. You’ll have an unique opportunity to meet many
+Python-enthusiastic developers, users and professionals.
+
+As a sponsor of EuroPython 2020, you will directly help to promote the work of a
+great open-source community and help further its development.
+
+**Your benefits as a EuroPython Sponsor**
+
+- **Reach out to more than 1200 inspired attendees**
+- Socializing: Make new contacts and network!
+- Targeting: Reach out directly to people with a real interest in the conference
+  topic and run a targeted marketing!
+- Image cultivation: Present yourself as a Python-related open source company!
+- Sponsoring: Support a great open source community!
+- Know-how transfer: Educate your own employees and benefit from the knowledge
+  of the community!
+- Head hunting: Take advantage of the conference for recruitment!
+- New business: Use the conference to attract new projects!
+
+**Sponsor Team**
+
+If you are interested in sponsoring EuroPython, please send an email to our
+sponsor team at:
+
+- sponsoring@europython.eu.
+
+**Sponsor Packages & Brochure**
+
+We are still preparing the EuroPython 2020 sponsor brochure. In the meantime,
+please have a look at our EuroPython 2019 Sponsor Brochure and the packages we
+had last year.
+
+- [EuroPython 2019 Sponsor Brochure](https://ep2019.europython.eu/sponsor/brochure/)
+- [EuroPython 2019 Sponsor Packages](https://ep2019.europython.eu/sponsor/packages/)
+
+The prices for 2020 haven’t been finalized yet, but will stay at similar levels.
+We will also offer a 10% discount for early bird sponsors.
