+++ pages/index.tsx
@@ -2,205 +2,19 @@ import { serialize } from "next-mdx-remote/serialize";
 import { MDXRemote } from "next-mdx-remote";
 
 import { promises as fs } from "fs";
-import Head from "next/head";
 import path from "path";
+import { Layout } from "../components/layout";
 
-const Header = () => (
-  <header>
-    <svg
-      viewBox="0 0 230.346 232.912"
-      width="230.346"
-      height="232.912"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g transform="matrix(1.789176, 0, 0, 1.789176, -133.051788, 0.002804)">
-        <g>
-          <g>
-            <g>
-              <path
-                style={{ fill: "#f18a00" }}
-                d="M156.1,108.08c3.22,0,5.84,2.63,5.84,5.89c0,3.28-2.62,5.94-5.84,5.94c-3.21,0-5.82-2.66-5.82-5.94      C150.28,110.71,152.89,108.08,156.1,108.08z M173.55,33.18v13.81c0,10.68-9.06,19.69-19.4,19.69h-31.02      c-7.31,0-13.53,5.37-15.13,12.29c-0.26,1.13-0.4,2.3-0.4,3.49v29.56c0,8.42,7.32,13.37,15.53,15.79      c9.83,2.89,19.25,3.41,31.02,0c7.82-2.27,15.53-6.81,15.53-15.79v-11.84h-31.02v-3.94h31.02h15.52c9.03,0,12.39-6.3,15.52-15.74      c3.26-9.73,3.11-19.08,0-31.56c-2.22-9-6.48-15.76-15.52-15.76H173.55z"
-              />
-            </g>
-          </g>
-          <path
-            style={{ fill: "#3dae2b" }}
-            d="M121.17,9.52c3.2,0,5.82,2.65,5.82,5.94c0,3.25-2.62,5.89-5.82,5.89c-3.22,0-5.83-2.64-5.83-5.89    C115.34,12.18,117.95,9.52,121.17,9.52z M137.97,0c-5.32,0.02-10.38,0.48-14.84,1.26c-13.14,2.33-15.53,7.19-15.53,16.15l0,11.83    l31.05-0.01l0,3.94l-31.05,0.01l-11.65,0c-9.03,0-16.93,5.43-19.4,15.77c-2.86,11.82-2.97,19.2,0.01,31.56    c2.21,9.2,7.48,15.74,16.51,15.74l10.68,0l0-14.18c0-10.25,8.86-19.3,19.4-19.3l31.02-0.01c8.63,0,15.53-7.12,15.52-15.78    l-0.01-29.57c0-8.42-7.1-14.73-15.53-16.15C148.8,0.37,143.27-0.03,137.97,0z"
-          />
-          <path
-            style={{ fill: "#006e43" }}
-            d="M169.68,24.62v-7.21c0-0.51-0.03-1-0.09-1.49c-0.01-0.08-0.02-0.17-0.03-0.25c-0.03-0.23-0.05-0.47-0.09-0.7    c-0.01,0-0.02,0-0.02,0c-1.21-7.19-7.71-12.45-15.31-13.72C148.98,0.4,143.65,0,138.51,0c-0.18,0-0.36,0-0.55,0    c-5.32,0.02-10.38,0.48-14.84,1.26c-1.39,0.25-2.65,0.52-3.81,0.83c-0.01,0-0.03,0.01-0.04,0.01    c-9.77,2.58-11.68,7.31-11.68,15.31l0,11.83l26.17-0.01c0,0,0,0,0,0h4.88l0,3.93l-25.4,0.01c0,0,0,0,0,0h-4.48c0,0,0,0,0,0    l-1.16,0l-11.65,0c-0.34,0-0.67,0.02-1,0.04c0,0-0.01,0-0.01-0.01c-0.05,0-0.09,0.01-0.14,0.01c-0.42,0.02-0.84,0.06-1.25,0.1    c-0.12,0.01-0.25,0.03-0.37,0.04c-0.4,0.05-0.8,0.12-1.2,0.19c-0.13,0.02-0.26,0.05-0.39,0.08c-0.38,0.08-0.76,0.17-1.13,0.27    c-0.14,0.04-0.29,0.08-0.43,0.12c-0.35,0.1-0.71,0.22-1.05,0.34c-0.16,0.06-0.31,0.11-0.46,0.17c-0.33,0.13-0.65,0.26-0.97,0.4    c-0.17,0.08-0.34,0.16-0.5,0.24c-0.3,0.14-0.59,0.29-0.88,0.45c-0.18,0.1-0.36,0.21-0.54,0.32c-0.26,0.16-0.53,0.32-0.78,0.49    c-0.19,0.13-0.38,0.27-0.57,0.4c-0.23,0.17-0.46,0.34-0.69,0.52c-0.2,0.16-0.4,0.33-0.6,0.51c-0.2,0.17-0.4,0.35-0.59,0.53    c-0.21,0.2-0.41,0.41-0.61,0.62c-0.17,0.18-0.34,0.35-0.5,0.53c-0.21,0.24-0.42,0.5-0.62,0.75c-0.14,0.17-0.28,0.34-0.41,0.52    c-0.22,0.29-0.42,0.6-0.62,0.91c-0.11,0.16-0.22,0.31-0.32,0.48c-0.22,0.35-0.42,0.72-0.62,1.1c-0.07,0.14-0.15,0.27-0.23,0.41    c-0.22,0.44-0.43,0.9-0.63,1.36c-0.04,0.09-0.08,0.17-0.12,0.26c-0.46,1.13-0.86,2.33-1.16,3.61c-0.18,0.73-0.34,1.43-0.49,2.13    c-0.02,0.08-0.03,0.16-0.05,0.24c-0.12,0.53-0.23,1.05-0.33,1.56c-0.06,0.3-0.12,0.59-0.17,0.89c-0.05,0.28-0.11,0.57-0.16,0.85    c-0.07,0.37-0.13,0.73-0.19,1.09c-0.01,0.08-0.03,0.16-0.04,0.24c0,0,0,0,0,0c-0.64,3.96-0.87,7.58-0.67,11.35c0,0,0,0,0,0    c0,0.08,0.01,0.16,0.02,0.24c0.02,0.4,0.05,0.81,0.08,1.22c0.02,0.2,0.03,0.4,0.05,0.6c0.03,0.37,0.07,0.74,0.11,1.11    c0.03,0.3,0.07,0.6,0.11,0.9c0.04,0.29,0.07,0.57,0.11,0.86c0.07,0.46,0.14,0.94,0.22,1.41c0.02,0.15,0.05,0.29,0.07,0.44    c0.01,0.06,0.02,0.11,0.03,0.16c0,0,0,0,0,0c0.27,1.57,0.61,3.22,1.01,4.96c0,0,0,0,0,0c0.1,0.42,0.19,0.84,0.29,1.27    c0.52,2.17,1.22,4.18,2.09,6c0.06,0.12,0.11,0.25,0.17,0.37c0.01,0.03,0.02,0.05,0.04,0.08c0,0,0,0,0.01,0    c0.36,0.72,0.76,1.4,1.18,2.06c0.01,0.02,0.01,0.04,0.02,0.05c0.06,0.1,0.13,0.18,0.19,0.28c0.09,0.13,0.17,0.25,0.26,0.38    c0.19,0.26,0.38,0.52,0.58,0.77c0.09,0.11,0.17,0.22,0.25,0.32c0.02,0.02,0.04,0.05,0.06,0.07c0.01,0,0.01,0.01,0.02,0.01    c2.01,2.4,4.56,4.09,7.68,4.87c0.17,0.04,0.34,0.09,0.52,0.13c0.3,0.06,0.61,0.11,0.91,0.16c0.1,0.02,0.2,0.04,0.31,0.05    c-0.01,0-0.02-0.01-0.03-0.01c0.72,0.09,1.47,0.16,2.24,0.16l10.68,0l0-14.18c0-10.25,8.86-19.3,19.4-19.3l15.8,0c0,0,0,0,0,0    h14.18c0,0,0,0,0-0.01l1.03,0c1.38,0,2.7-0.2,3.97-0.54c0,0.01,0.01,0.01,0.01,0.02c0.13-0.04,0.25-0.09,0.38-0.13    c0.23-0.07,0.45-0.14,0.67-0.22c0.27-0.09,0.54-0.19,0.8-0.3c0.23-0.1,0.46-0.2,0.68-0.31c0.17-0.08,0.34-0.14,0.5-0.23    c-0.01,0-0.01-0.01-0.02-0.01c5.07-2.61,8.53-7.95,8.52-14.07l0-11.18c0,0,0,0,0,0v-8.74c0,0,0,0,0,0L169.68,24.62    C169.68,24.62,169.68,24.62,169.68,24.62z M122.53,21.19c-0.19,0.05-0.39,0.08-0.59,0.11c-0.03,0-0.06,0.01-0.09,0.01    c-0.22,0.03-0.44,0.04-0.67,0.04c0,0-0.01,0-0.01,0c-0.03,0-0.06,0-0.09,0c-0.16,0-0.31-0.01-0.46-0.02    c-0.09-0.01-0.18-0.02-0.27-0.04c-0.09-0.01-0.18-0.02-0.27-0.04c-0.1-0.02-0.2-0.04-0.3-0.07c-0.07-0.02-0.14-0.04-0.21-0.06    c-0.11-0.03-0.21-0.07-0.32-0.1c-0.06-0.02-0.11-0.04-0.17-0.06c-0.11-0.04-0.23-0.09-0.34-0.14c-0.04-0.02-0.08-0.04-0.12-0.06    c-0.12-0.06-0.24-0.12-0.36-0.19c-0.01-0.01-0.03-0.02-0.04-0.03c-1.42-0.85-2.45-2.28-2.77-3.97c0,0-0.01-0.01-0.01-0.01    c-0.05-0.27-0.08-0.54-0.09-0.81c0,0,0.01,0,0.01,0c0-0.1-0.01-0.19-0.01-0.29c0-3.28,2.61-5.94,5.82-5.94    c0.04,0,0.08,0.01,0.13,0.01c0.16,0,0.33,0.01,0.49,0.03c0.1,0.01,0.19,0.03,0.28,0.04c0.1,0.02,0.21,0.03,0.31,0.05    c0.1,0.02,0.2,0.05,0.3,0.08c0.09,0.03,0.19,0.05,0.28,0.08c0.1,0.03,0.2,0.07,0.29,0.11c0.09,0.03,0.18,0.07,0.26,0.11    c0.09,0.04,0.19,0.09,0.28,0.14c0.09,0.04,0.17,0.09,0.25,0.14c0.09,0.05,0.17,0.1,0.26,0.16c0.08,0.05,0.16,0.11,0.24,0.17    c0.08,0.06,0.16,0.12,0.24,0.18c0.08,0.06,0.15,0.13,0.22,0.19c0.07,0.07,0.15,0.13,0.22,0.2c0.07,0.07,0.13,0.14,0.2,0.21    c0.97,1.06,1.57,2.48,1.57,4.04C126.99,18.23,125.08,20.56,122.53,21.19z M76.38,74.9c0.81,0.41,1.67,0.73,2.53,1.01    c0.93,0.3,1.87,0.56,2.81,0.78c-0.91,0.26-1.82,0.55-2.71,0.91c-0.66,0.26-1.3,0.56-1.88,0.94C76.84,77.29,76.59,76.08,76.38,74.9    z M154.15,61.75l-0.77,0c0.36-1.35,0.67-2.71,0.92-4.09c0.81,0.82,2.15,2.26,3.2,3.68C156.43,61.6,155.31,61.75,154.15,61.75z     M168.68,27.64c-0.41,0.28-0.79,0.6-1.12,0.99c-1.1,1.34-1.36,2.85-0.76,4.46c0.38,1.04,1.02,1.79,1.88,2.28l0,11.6    c0,6-3.54,11.17-8.6,13.48c-1.69-1.07-3.65-2.9-4.67-3.9c2.57-0.47,5.11-1.04,7.55-2.02c1.16-0.47,2.3-1.01,3.12-2    c1.1-1.34,1.36-2.85,0.76-4.46c-0.59-1.59-1.78-2.53-3.45-2.84c-0.38-0.07-0.77-0.07-1.16-0.1c0-0.09,0-0.17,0-0.25    c0.06-2.02-1.18-3.77-3.11-4.41c-1.87-0.62-3.95,0.11-5.09,1.74c-0.82,1.17-1.31,2.48-1.75,3.82c-0.12,0.37-0.22,0.74-0.33,1.11    c-0.36-1.49-0.79-2.96-1.37-4.39c-0.47-1.16-1.01-2.3-2-3.12c-1.34-1.1-2.85-1.36-4.46-0.76c-1.59,0.59-2.53,1.78-2.84,3.45    c-0.07,0.38-0.07,0.77-0.1,1.16c-0.09,0-0.17,0-0.25-0.01c-2.02-0.06-3.77,1.18-4.41,3.11c-0.62,1.87,0.11,3.95,1.74,5.09    c1.17,0.82,2.48,1.31,3.82,1.75c0.93,0.3,1.87,0.56,2.81,0.78c-0.91,0.26-1.82,0.55-2.71,0.91c-1.16,0.47-2.3,1.01-3.12,2    c-1.1,1.34-1.36,2.85-0.76,4.46c0.02,0.06,0.05,0.11,0.08,0.17l-15.25,0c-5.34,0-10.44,2.17-14.36,6.1    c-3.84,3.85-6.04,9.03-6.04,14.2l0,13.18l-9.68,0c-2.94,0-6.36-0.76-9.36-3.4c1.62,0.19,3.26-0.54,4.23-1.92    c0.82-1.17,1.31-2.48,1.75-3.82c0.64-1.95,1.08-3.94,1.44-5.96c1.18,1.2,3.54,3.73,4.4,5.56c1.01,2.14,0.94,2.48,0.94,2.48    l1.9-1.9l0,0l1.9-1.9c0,0-0.33,0.07-2.48-0.94c-1.83-0.86-4.36-3.22-5.56-4.4c2.57-0.47,5.11-1.04,7.55-2.02    c1.16-0.47,2.3-1.01,3.12-2c1.1-1.34,1.36-2.85,0.76-4.46c-0.59-1.59-1.78-2.53-3.45-2.84c-0.38-0.07-0.77-0.07-1.16-0.1    c0-0.09,0-0.17,0-0.25c0.06-2.02-1.18-3.77-3.11-4.41c-1.87-0.62-3.95,0.11-5.09,1.74c-0.82,1.17-1.31,2.48-1.75,3.82    c-0.12,0.37-0.22,0.74-0.33,1.11c-0.36-1.49-0.79-2.96-1.36-4.39c-0.47-1.16-1.01-2.3-2-3.12c-1.34-1.1-2.85-1.36-4.46-0.76    c-1.59,0.59-2.53,1.78-2.84,3.45c-0.07,0.38-0.07,0.77-0.1,1.16c-0.09,0-0.17,0-0.25-0.01c-0.86-0.02-1.67,0.19-2.37,0.6    c-0.13-3.26,0.09-6.46,0.62-9.95c0.26,0.15,0.53,0.27,0.82,0.37c1.87,0.62,3.95-0.11,5.09-1.74c0.82-1.17,1.31-2.48,1.75-3.82    c0.64-1.95,1.08-3.94,1.44-5.96c1.18,1.2,3.54,3.73,4.4,5.56c1.01,2.14,0.94,2.48,0.94,2.48l1.9-1.9l0,0l1.9-1.9    c0,0-0.33,0.07-2.48-0.94c-1.83-0.86-4.36-3.22-5.56-4.4c2.57-0.47,5.11-1.04,7.55-2.02c1.16-0.47,2.3-1.01,3.12-2    c1.1-1.34,1.36-2.85,0.76-4.46c-0.25-0.69-0.62-1.25-1.09-1.69l11,0c-0.49,0.58-0.83,1.29-0.99,2.14    c-0.07,0.38-0.07,0.77-0.1,1.16c-0.09,0-0.17,0-0.25-0.01c-2.02-0.06-3.77,1.18-4.41,3.11c-0.62,1.87,0.11,3.95,1.74,5.09    c1.17,0.82,2.48,1.31,3.82,1.75c0.93,0.3,1.87,0.56,2.81,0.78c-0.91,0.26-1.82,0.55-2.71,0.91c-1.16,0.47-2.3,1.01-3.12,2    c-1.1,1.34-1.36,2.85-0.76,4.46c0.59,1.59,1.78,2.53,3.45,2.84c0.38,0.07,0.77,0.07,1.16,0.09c0,0.09,0,0.17-0.01,0.25    c-0.06,2.02,1.18,3.77,3.11,4.41c1.87,0.62,3.95-0.11,5.09-1.74c0.82-1.17,1.31-2.48,1.75-3.82c0.64-1.95,1.08-3.94,1.44-5.96    c1.18,1.2,3.54,3.73,4.4,5.56c1.01,2.14,0.94,2.48,0.94,2.48l1.9-1.9l0,0l1.9-1.9c0,0-0.33,0.07-2.48-0.94    c-1.83-0.86-4.36-3.22-5.56-4.4c2.57-0.47,5.11-1.04,7.55-2.02c1.16-0.47,2.3-1.01,3.12-2c1.1-1.34,1.36-2.85,0.76-4.46    c-0.59-1.59-1.78-2.53-3.45-2.84c-0.38-0.07-0.77-0.07-1.16-0.1c0-0.09,0-0.17,0.01-0.25c0.06-2.02-1.18-3.77-3.11-4.41    c-1.87-0.62-3.95,0.11-5.09,1.74c-0.82,1.17-1.31,2.48-1.75,3.82c-0.12,0.37-0.22,0.74-0.33,1.11c-0.36-1.49-0.79-2.96-1.36-4.39    c-0.37-0.93-0.8-1.83-1.46-2.57l24.21-0.01l0.12,0c0.32,0.21,0.67,0.39,1.05,0.51c1.87,0.62,3.95-0.11,5.09-1.74    c0.82-1.17,1.31-2.48,1.75-3.82c0.64-1.95,1.08-3.94,1.44-5.96c1.18,1.2,3.54,3.73,4.4,5.56c1.01,2.15,0.94,2.48,0.94,2.48    l1.9-1.9l0,0l1.9-1.9c0,0-0.33,0.07-2.48-0.94c-1.83-0.86-4.36-3.22-5.56-4.4c2.57-0.47,5.11-1.04,7.55-2.02    c1.16-0.47,2.3-1.01,3.12-2c1.1-1.34,1.36-2.85,0.76-4.46c-0.59-1.59-1.78-2.53-3.45-2.84c-0.38-0.07-0.77-0.07-1.16-0.1    c0-0.09,0-0.17,0-0.25c0.06-2.02-1.18-3.77-3.11-4.41c-1.87-0.62-3.95,0.11-5.09,1.74c-0.82,1.17-1.31,2.48-1.75,3.82    c-0.12,0.37-0.22,0.74-0.33,1.11c-0.36-1.49-0.79-2.96-1.37-4.39c-0.47-1.16-1.01-2.3-2-3.12c-1.34-1.1-2.85-1.36-4.46-0.76    c-1.59,0.59-2.53,1.78-2.84,3.45c-0.07,0.38-0.07,0.77-0.1,1.16c-0.09,0-0.17,0-0.25-0.01c-2.02-0.06-3.77,1.18-4.41,3.11    c-0.62,1.87,0.11,3.95,1.74,5.09c1.17,0.82,2.48,1.31,3.82,1.75c0.93,0.3,1.87,0.56,2.81,0.78c-0.91,0.26-1.82,0.55-2.71,0.91    c-1.16,0.47-2.3,1.01-3.12,2c-1.1,1.34-1.36,2.85-0.76,4.46c0.16,0.43,0.37,0.81,0.62,1.14l-24.11,0.01    c0.7-0.33,1.34-0.83,1.8-1.5c0.82-1.17,1.31-2.48,1.75-3.82c0.64-1.95,1.08-3.94,1.44-5.96c0.38,0.39,0.9,0.93,1.44,1.54    c0.47,0.95,1.14,1.77,1.97,2.41c0.4,0.56,0.75,1.11,0.99,1.61c1.01,2.14,0.94,2.48,0.94,2.48l1.9-1.9l0,0l0.78-0.78    c3.54-0.25,6.35-3.23,6.35-6.87c0-1.85-0.72-3.53-1.88-4.77c0.46-1.03,0.48-2.14,0.05-3.31c-0.59-1.59-1.78-2.53-3.45-2.84    c-0.38-0.07-0.77-0.07-1.16-0.1c0-0.09,0-0.17,0-0.25c0.02-0.55-0.08-1.07-0.25-1.57c0.63-0.14,1.29-0.27,1.99-0.39    c4.52-0.79,9.59-1.22,14.68-1.25l0.54,0c5.26,0,10.47,0.42,15.46,1.24c6.6,1.1,13.25,5.79,14.48,12.83    c-1.59,0.3-2.9,1.42-3.43,3.02c-0.62,1.87,0.11,3.95,1.74,5.09c0.6,0.42,1.25,0.75,1.91,1.04L168.68,27.64z"
-          />
-          <path
-            style={{ fill: "#ffcfa1" }}
-            d="M201.61,72.66c-0.26,1.71-0.63,3.39-1.1,5.08c-0.95,1.22-2.4,1.91-3.88,2.2c-1.42,0.27-2.81,0.16-4.11-0.54    c-2.43-1.3-2.76-3.9-1.01-5.87c1.81-2.03,4.14-2.65,6.79-2.08c0.07,0.02,0.14,0.03,0.22,0.04c0.05,0.01,0.11,0.01,0.2,0.03    c0-5.7,0-11.4,0-17.12c-2.05,0.37-4.09,0.74-6.12,1.11c-2.92,0.54-5.83,1.1-8.75,1.6c-0.49,0.08-0.52,0.29-0.52,0.68    c0,6.59-0.02,13.18-0.03,19.77c-0.01,3.23-1.81,5.31-5.03,5.82c-0.08,0.01-2.72,0.1-4.33-0.78c-0.88-0.48-1.47-1.22-1.64-2.27    c-0.21-1.3,0.19-2.41,1.05-3.33c1.83-1.95,4.09-2.59,6.71-2.02c0.12,0.03,0.25,0.03,0.41,0.05c0-8.43,0-16.84,0-25.29    c6.19-1,12.37-2.01,18.59-3.02c0.25,0.79,0.48,1.6,0.7,2.47c0.83,3.33,1.44,6.39,1.83,9.29C201.58,63.2,201.58,67.93,201.61,72.66    z M138.63,67.67c2.19,3.39,8.45,14.03,10.69,27.57h-6.53c-0.25-1.32-0.58-2.65-0.99-4.02c-2.27-7.55-4.69-12.15-6.55-14.86v45.54    l1.9,3.26l0.59-0.89c0.08-0.12,6.31-9.71,5.74-23.09h6.53c0.22,3.38,0.15,6.9-0.32,10.48c-0.26,0.73-0.41,1.5-0.41,2.32    c0,0.07,0.01,0.14,0.01,0.21c-0.44,2.42-1.05,4.86-1.91,7.31c-0.96,2.75-1.96,5.18-2.99,7.32c-1.98,0.24-3.93,0.36-5.86,0.36    c-4.92,0-9.73-0.74-15.12-2.33c-2.26-0.67-5.14-1.74-7.77-3.4c-1.28-3.59-2.84-8.03-4.75-13.53c-0.77-2.23-1.55-4.45-2.3-6.62    V82.46c0-1.1,0.13-2.2,0.37-3.27c1.09-4.69,4.35-8.46,8.5-10.3c0.05,0.05,0.11,0.11,0.17,0.16c0.93,0.9,1.61,1.56,3.27,1.29    c1.08-0.17,2.43-1.23,3.83-2.66H138.63z M114.09,76.38c-1.44-0.67-2.19-1.24-2.67-1.61c-0.04-0.03-0.08-0.06-0.12-0.09v6.06    l2.8,4.81V76.38z M118.88,77.88c-0.53-0.04-1.06-0.13-1.56-0.29c-0.44-0.14-0.85-0.28-1.23-0.41v11.78l2.8,4.8V77.88z     M123.68,77.1c-0.92,0.32-1.86,0.57-2.8,0.7V97.2l2.8,4.8V77.1z M128.47,74.78c-0.82,0.5-1.77,1.02-2.8,1.5v29.15l2.8,4.8V74.78z     M133.26,73.84c-0.68-0.71-1.1-1-1.16-1.04l-0.4-0.27l-0.38,0.3c-0.02,0.01-0.33,0.26-0.85,0.63v40.2l2.8,4.81V73.84z     M159.63,85.15c0,0.19,0,0.39,0,0.51c-0.86-0.04-1.67-0.16-2.47-0.09c-1.88,0.18-3.44,0.98-4.47,2.63c-1,1.61-0.51,3.55,1.13,4.48    c1.25,0.7,2.58,0.83,3.98,0.6c2.71-0.46,4.37-2.34,4.37-5.07c0.02-6.8,0.03-13.6,0.05-20.4c0-0.18,0-0.36,0-0.64    c4.42,0.63,6.98,3.61,9.34,6.9c-0.2-2.39-1.21-4.4-3.04-5.95c-1.15-0.97-2.42-1.8-3.61-2.73c-0.23-0.18-0.45-0.36-0.68-0.54    c-1.44,0.86-2.99,1.54-4.62,2.02C159.63,72.97,159.63,79.06,159.63,85.15z"
-          />
-        </g>
-      </g>
-    </svg>
-
-    <h1>EuroPython 2022</h1>
-    <p>July 11-17th 2022; Dublin Ireland &amp; Remote</p>
-  </header>
-);
-
-const Footer = () => (
-  <footer>
-    <div className="container">
-      <div className="row">
-        <div className="col-md-12">
-          <a href="https://www.europython-society.org">
-            <img src="/epslogo.png" />
-          </a>
-          <p
-            style={{
-              marginTop: "1em",
-              fontSize: "0.9em",
-            }}
-            className="text-light"
-          >
-            EuroPython Society (EPS)
-            <br />
-            Ramnebacken 45
-            <br />
-            424 38 Agnesberg
-            <br />
-            Sweden
-          </p>
-          <p>
-            <a href="https://www.europython-society.org" target="_blank">
-              EPS website
-            </a>
-          </p>
-        </div>
-      </div>
-
-      <div
-        className="row text-center"
-        style={{
-          marginTop: "3em",
-          fontFamily: "monospace",
-        }}
-      >
-        <div className="col-md-12">
-          <p>
-            <a href="https://blog.europython.eu" target="_blank">
-              blog.europython.eu
-            </a>
-          </p>
-          <p>
-            <a
-              href="https://twitter.com/europython"
-              target="_blank"
-              className="btn-outline-primary"
-            >
-              twitter.com/europython
-            </a>
-          </p>
-          <p>
-            <a
-              href="https://www.linkedin.com/company/europython"
-              target="_blank"
-              className="btn-outline-primary"
-            >
-              linkedin.com/company/europython
-            </a>
-          </p>
-        </div>
-      </div>
-
-      <div>
-        <p>
-          Copyright &copy; 2022, EuroPython Society,{" "}
-          <a href="mailto:info@europython-society.org">
-            info@europython-society.org
-          </a>
-        </p>
-
-        <hr />
-
-        <a href="https://vercel.com/?utm_source=europython-2022&utm_campaign=oss">
-          <img
-            src="https://www.datocms-assets.com/31049/1618983297-powered-by-vercel.svg"
-            alt="powered by vercel"
-          />
-        </a>
-      </div>
-    </div>
-  </footer>
-);
-
-export default function DocumentPage({ source }: { source: any }) {
+export default function IndexPage({ source }: { source: any }) {
   return (
-    <>
-      <Head>
-        <meta charSet="utf-8" />
-        <meta
-          name="viewport"
-          content="width=device-width, initial-scale=1, shrink-to-fit=no"
-        />
-
-        <meta http-equiv="cleartype" content="on" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-
-        <link
-          rel="alternate"
-          type="application/rss+xml"
-          title="EuroPython Blog RSS Feed"
-          href="http://blog.europython.eu/rss"
-        />
-
-        <link rel="shortcut icon" type="image/png" href="/favicon.png" />
-
-        <title>
-          EuroPython 2022 | July 11th-17th 2022 | Dublin Ireland & Remote
-        </title>
-        <meta
-          name="title"
-          content="EuroPython 2022 | July 11th-17th 2022 | Dublin Ireland & Remote"
-        />
-        <meta name="description" content="" />
-        <meta name="author" content="EuroPython" />
-
-        <meta property="og:type" content="website" />
-        <meta property="og:url" content="https://ep2022.europython.eu/" />
-        <meta
-          property="og:title"
-          content="EuroPython 2022 | July 11th-17th 2022 | Dublin Ireland & Remote"
-        />
-        <meta property="og:description" content="" />
-        <meta
-          property="og:image"
-          content="https://ep2022.europython.eu/2022-europython-social-card.png"
-        />
-
-        <meta property="twitter:card" content="summary_large_image" />
-        <meta property="twitter:url" content="https://ep2022.europython.eu/" />
-        <meta
-          property="twitter:title"
-          content="EuroPython 2022 | July 11th-17th 2022 | Dublin Ireland & Remote"
-        />
-        <meta property="twitter:description" content="" />
-        <meta
-          property="twitter:image"
-          content="https://ep2022.europython.eu/2022-europython-social-card.png"
-        />
-
-        <script
-          defer
-          data-domain="ep2022.europython.eu"
-          src="https://plausible.io/js/plausible.js"
-        ></script>
-      </Head>
-      <Header />
+    <Layout>
       <MDXRemote {...source} />
-      <Footer />
-    </>
+    </Layout>
   );
 }
 
 export async function getStaticProps() {
-  const faqPath = path.join(process.cwd(), 'data/faq.md')
+  const faqPath = path.join(process.cwd(), "data/faq.md");
 
   const content = await fs.readFile(faqPath);
   const mdxSource = await serialize(content.toString());
